Initial splitting of the behavior in kitconcept.volto.footer
