+++ frontend/packages/volto-light-theme/src/components/Footer/slots/Colophon.tsx
@@ -2,7 +2,7 @@ import { Container } from '@plone/components';
 import Copyright from './Copyright';
 import Logo from '../../Logo/Logo';
 import type { Content } from '@plone/types';
-import { useLiveData } from '@kitconcept/volto-light-theme/helpers/liveData';
+import { useLiveData } from '@kitconcept/volto-light-theme/helpers/useLiveData';
 import type { SiteHeaderSettings } from '../../../types';
 
 const Colophon = ({ content }: { content: Content }) => {
@@ -11,7 +11,7 @@ const Colophon = ({ content }: { content: Content }) => {
     'voltolighttheme.header',
     'logo',
   );
-  console.log(logo);
+
   return (
     <Container className="colophon">
       <div className="powered-by">Powered by Plone & Light Theme</div>
