+++ frontend/packages/volto-light-theme/src/config/slots.ts
@@ -3,8 +3,13 @@ import Theming from '../components/Theming/Theming';
 import FooterLogos from '../components/Footer/slots/FooterLogos';
 import FollowUsLogoAndLinks from '../components/Footer/slots/FollowUsLogoAndLinks';
 import Colophon from '../components/Footer/slots/Colophon';
-import Copyright from '../components/Footer/slots/Copyright';
 import CoreFooter from '../components/Footer/slots/CoreFooter';
+import type { Content } from '@plone/types';
+
+export function hasInheritedBehavior(behavior: string) {
+  return ({ content }: { content: Content }) =>
+    Object.keys(content?.['@components']?.inherit || {}).includes(behavior);
+}
 
 export default function install(config: ConfigType) {
   config.registerSlotComponent({
@@ -19,11 +24,12 @@ export default function install(config: ConfigType) {
     component: FooterLogos,
   });
 
-  // config.registerSlotComponent({
-  //   name: 'coreFooter',
-  //   slot: 'footer',
-  //   component: CoreFooter,
-  // });
+  config.registerSlotComponent({
+    name: 'coreFooter',
+    slot: 'footer',
+    component: CoreFooter,
+    predicates: [hasInheritedBehavior('kitconcept.volto.footer')],
+  });
 
   config.registerSlotComponent({
     name: 'PostFooterFollowUsLogoAndLinks',
