+++ frontend/packages/volto-light-theme/src/types.d.ts
@@ -61,7 +61,6 @@ export type SiteThemeSettings = {
 };
 
 export type SiteFooterSettings = {
-  has_enhanced_footer: boolean;
   footer_logos: Array<footerLogo>;
   footer_logos_container_width: string;
   footer_logos_size: string;
@@ -72,7 +71,6 @@ export type SiteFooterSettings = {
   footer_column_middle: Array<Link>;
   footer_column_right_header: string;
   footer_column_right: Array<Link>;
-  followus_links: Array<iconLink>;
   footer_logo: Image & { data?: string };
   footer_links: Array<Link>;
 };
@@ -88,7 +86,6 @@ export type CustomInheritBehavior<T> = {
 declare module '@plone/types' {
   export interface Content {
     footer_logos: Array<footerLogo>;
-    followus_links: Array<iconLink>;
     footer_links: Array<Link>;
   }
 }
